backend/app/config.py
"""Application configuration."""

import os
from dotenv import load_dotenv

load_dotenv()


class Config:
    """Base configuration for Flask app."""

    FRONTEND_URL = os.getenv("FRONTEND_URL", "http://localhost:5173")
    FIREBASE_CREDENTIALS_PATH = os.getenv(
        "FIREBASE_CREDENTIALS_PATH", "firebase-service-account.json"
    )
